+++ src/elements/db/ItemQuery.php
@@ -120,6 +120,10 @@ class ItemQuery extends ElementQuery
             'element_item_sites.title AS elementTitle',
         ]);
 
+        // Join the linked-to element's content
+        $this->subQuery->innerJoin('{{%content}} element_content', '[[wishlist_items.elementId]] = [[element_content.elementId]] AND [[wishlist_items.elementSiteId]] = [[element_content.siteId]]');
+        $this->subQuery->addSelect('element_content.title AS elementTitle');
+
         if ($this->id) {
             $this->subQuery->andWhere(Db::parseParam('wishlist_items.id', $this->id));
         }